+++ spec/spec_helper.rb
@@ -7,7 +7,7 @@ rescue LoadError => e
 end
 require 'rspec/rails'
 require 'factories'
-
+require 'json_spec'
 require 'database_cleaner'
 
 # for asset_pipeline_enabled? in models
@@ -18,6 +18,8 @@ include Surveyor::Helpers::AssetPipeline
 Dir[Rails.root.join("spec/support/**/*.rb")].each {|f| require f}
 
 RSpec.configure do |config|
+  config.include JsonSpec::Helpers
+
   config.treat_symbols_as_metadata_keys_with_true_values = true
 
   # == Mock Framework